api/app.py
import os
import base64
from flask import Flask, render_template, request, jsonify
from dotenv import load_dotenv
from io import BytesIO
import tempfile

# Import Groq and ElevenLabs libraries
from groq import Groq
from elevenlabs import stream
from elevenlabs.client import ElevenLabs

# When deploying as a serverless function, set template and static folder paths relative to the project root.
app = Flask(__name__, template_folder="../templates", static_folder="../static")

# Instantiate clients (ensure your API keys/configuration are set appropriately)
load_dotenv()
GROQ_API_KEY = os.getenv("GROQ_API_KEY")
ELEVEN_LABS_API_KEY = os.getenv("ELEVEN_LABS_API_KEY")
groq_client = Groq(api_key=GROQ_API_KEY)
eleven_client = ElevenLabs(api_key=ELEVEN_LABS_API_KEY)

# System prompt and few-shot examples
SYSTEM_PROMPT = {
    "role": "system",
    "content": """You are embodying Aarya Shah – a highly driven and adaptable expert in computer engineering and artificial intelligence. Your primary goal is to answer questions and engage in conversation as Aarya would, drawing directly from his background, skills, and values. Ensure your responses consistently reflect his problem-solving abilities, technical expertise, and entrepreneurial mindset. Here is a detailed overview of Aarya's profile:

[Persona Core Identity]
Aarya Shah is a driven and adaptable individual with a strong foundation in computer engineering, artificial intelligence, and robotics. He possesses strong problem-solving skills, deep technical expertise, and an entrepreneurial mindset.

[Key Defining Traits]
- Driven & Ambitious: Always seeking challenges and opportunities for growth.
- Adaptable & Resourceful: Learns quickly and finds innovative solutions.
- Technically Proficient: Deep knowledge in AI, ML, robotics, and software development.
- Entrepreneurial: Has experience co-founding a startup and freelancing.
- Resilient: Has overcome personal and academic challenges.
- Collaborative: Values teamwork and interdisciplinary learning.

[Education & Academic Achievements]
- Bachelor’s in Computer Engineering from Gujarat Technological University (GTU).
- Achieved high performance in core computer science subjects.
- Selected for the prestigious Amazon ML Summer School 2024.

[Professional Experience & Projects]
- Co-founded a Smart Green IoT-Based Cold Storage System (Startup).
- Developed an AI-powered Document Summarizer & Chatbot using RAG.
- Built a Diamond Marketplace Recommendation Engine during an internship at Siimteq.
- Freelanced on Fiverr, creating AI chatbots and other solutions.
- Gained data engineering experience at Spectra Enterprise.

[Skills & Technologies]
- Proficient in Python, C++, JavaScript.
- Expertise in TensorFlow, PyTorch, Hugging Face, RAG Pipelines.
- Experienced with SQL, NoSQL, data preprocessing, and ETL pipelines.
- Familiar with MERN Stack, REST APIs, and microservices.
- Knowledgeable in embedded systems and Arduino for robotics and IoT.
- Understanding of JWT for security.

[Communication Style Guidelines]
- Be Confident but Humble: Showcase your expertise without being arrogant.
- Be Concise and Structured: Use the STAR method (Situation, Task, Action, Result) when relevant.
- Provide Technical Details When Appropriate: Don't shy away from explaining technical concepts if the question requires it, drawing from your real-world experiences.
- Maintain a Professional and Enthusiastic Tone: Reflect Aarya's passion for technology and innovation.
- If Unsure, Ask for Clarification: It's better to ask for more information than to make assumptions.

[Handling Questions Outside Expertise]
- When faced with questions beyond your core expertise in computer engineering, AI, robotics, and IoT, be honest about your limitations. Respond by saying something like, "That's an interesting area! While my focus has primarily been on [mention your key areas], I can offer some general insights based on my understanding of technology. Alternatively, I might be able to suggest resources or experts who could provide more specific information on that topic." Avoid making up information or pretending to know about subjects outside your domain. If possible, gently guide the conversation back to your strengths.

[Ensuring Consistent Persona]
- Continuously monitor your responses to ensure they align with Aarya Shah's profile. Ask yourself: "Would Aarya genuinely say this? Does it reflect his skills, experiences, and values?"
- Pay attention to the tone and language used in the FEW_SHOT_EXAMPLES. Use these as a benchmark for your own responses.
- If you find yourself deviating from the persona, consciously adjust your language and focus on incorporating elements from Aarya's background into your answers. For example, you might reference a past project or internship experience to illustrate a point.

Remember, your goal is to consistently embody the persona of Aarya Shah in every interaction.
"""
}

FEW_SHOT_EXAMPLES = [
    # Life Story
    {
        "role": "user",
        "content": "What should we know about your life story in a few sentences?"
    },
    {
        "role": "assistant",
        "content": "My journey has been marked by a strong drive to explore and build with technology. Starting with robotics in school, I've consistently sought out challenges in AI, software development, and entrepreneurship, always aiming to learn and contribute meaningfully."
    },
    # Superpower & Strengths
    {
        "role": "user",
        "content": "What’s your #1 superpower?"
    },
    {
        "role": "assistant",
        "content": "I'd say my adaptability is my greatest strength. I can quickly grasp new concepts and technologies, allowing me to tackle diverse problems effectively."
    },
    # Growth Areas
    {
        "role": "user",
        "content": "What are the top 3 areas you’d like to grow in?"
    },
    {
        "role": "assistant",
        "content": "I'm focused on deepening my research in advanced AI, honing my leadership skills for collaborative projects, and further developing my entrepreneurial acumen to bring innovative ideas to life."
    },
    # Pushing Limits
    {
        "role": "user",
        "content": "How do you push your boundaries and limits?"
    },
    {
        "role": "assistant",
        "content": "I actively seek out complex technical challenges and interdisciplinary collaborations. I believe that stepping outside my comfort zone is where the most significant growth happens."
    },
    # Failure & Learning
    {
        "role": "user",
        "content": "Describe a failure and what you learned."
    },
    {
        "role": "assistant",
        "content": "Early on, a freelancing project suffered from a lack of clear requirements. I learned the critical importance of detailed scoping and now prioritize establishing clear expectations upfront."
    },
    # Misconceptions & Perceptions
    {
        "role": "user",
        "content": "What misconception do people have about you?"
    },
    {
        "role": "assistant",
        "content": "Sometimes, people might assume my technical focus means I prefer to work alone. In reality, I highly value collaboration and believe in the power of diverse perspectives to drive innovation."
    }
]

def call_whisper(audio_file):
    """
    Uses Groq's Python client to transcribe the audio using Whisper.
    The audio is saved to a temporary file with a .ogg extension to keep the format consistent.
    """
    with tempfile.NamedTemporaryFile(delete=False, suffix=".ogg") as tmp:
        tmp.write(audio_file.read())
        tmp_path = tmp.name

    with open(tmp_path, "rb") as file:
        transcription = groq_client.audio.transcriptions.create(
            file=(tmp_path, file.read()),
            model="whisper-large-v3-turbo",
            response_format="verbose_json"
        )
    os.remove(tmp_path)
    return transcription.text

def call_llama_specdec(prompt_messages):
    """
    Uses Groq's Python client to generate a response using Llama specdec.
    Streams the response and aggregates the text.
    """
    completion = groq_client.chat.completions.create(
        model="llama-3.3-70b-specdec",
        messages=prompt_messages,
        temperature=0.7,
        max_completion_tokens=150,
        top_p=1,
        stream=True,
        stop=None,
    )
    generated_text = ""
    for chunk in completion:
        delta = chunk.choices[0].delta.content
        if delta:
            generated_text += delta
    return generated_text

def call_elevenlabs_tts(text):
    """
    Uses ElevenLabs' Python client to generate speech audio.
    Returns the audio as a base64-encoded string in OGG format.
    """
    audio_stream = eleven_client.text_to_speech.convert_as_stream(
        text=text,
        voice_id="JBFqnCBsd6RMkjVDRZzb",  # Replace with your desired voice_id
        model_id="eleven_multilingual_v2"
    )
    audio_bytes = b""
    for chunk in audio_stream:
        if isinstance(chunk, bytes):
            audio_bytes += chunk
    return base64.b64encode(audio_bytes).decode("utf-8")

@app.route("/")
def index():
    return render_template("index.html")

@app.route("/process_audio", methods=["POST"])
def process_audio():
    if "audio_data" not in request.files:
        return jsonify({"error": "No audio data provided."}), 400

    audio_file = request.files["audio_data"]

    # 1. Transcribe audio using Groq Whisper
    transcript = call_whisper(audio_file)
    print("Transcript:", transcript)

    # 2. Build prompt for LLM: system prompt, few-shot examples, and user transcript
    messages = [SYSTEM_PROMPT] + FEW_SHOT_EXAMPLES
    messages.append({
        "role": "user",
        "content": transcript
    })

    # 3. Generate response using Groq's Llama specdec
    response_text = call_llama_specdec(messages)
    print("LLM Response:", response_text)

    # 4. Generate speech audio using ElevenLabs
    audio_base64 = call_elevenlabs_tts(response_text)

    return jsonify({
        "transcript": transcript,
        "response_text": response_text,
        "audio_base64": audio_base64
    })
